Centralise app name config, and add a service account
Introduced APPLICATION_NAME files to each project to centralise the
configuration of the application name. This means there's only one
place to change.

Adds an explicit service account to each application. This will be
required if anyone wants to grant any permissions, so lket's get this
done up front for people.