Incorporate feedback from user testing!
Clarified some things, and impoved the deploy scripts. They're a bit
more robust and give much better user feedback.

The APPLICATION_NAME no longer clashes with the filename, and the
instructions for changing the name have been improved. We also now
check this from the deploy scripts so people won't overwrite the
template.