Add information about application secrets
We add the CAPI_API_KEY as an example secret, since this is one that
will be commonly used. Also documents the process for adding a new
secret.

For simplicity, we switch to using a shared service account that is
pre-granted access to all the secrets in Secret Manager.

Also demonstrate best practices by escaping the value obtained from
the `name` GET variable before displaying it.